+++ CHANGELOG.md
@@ -14,6 +14,7 @@ FEATURES
 * [gaiacli] You can now attach a simple text-only memo to any transaction, with the `--memo` flag
 * [lcd] Queried TXs now include the tx hash to identify each tx
 * [mockapp] CompleteSetup() no longer takes a testing parameter
+* Add benchmarking folder
 
 FIXES
 * \#1259 - fix bug where certain tests that could have a nil pointer in defer